+++ source/js/common/inject-react/join-us.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import JoinUs from "../components/join/join.jsx";
+import JoinUs from "../../components/join/join.jsx";
 
 /**
  * Inject newsletter signup forms
@@ -8,7 +8,7 @@ import JoinUs from "../components/join/join.jsx";
  * @param {String} siteUrl Foundation site base URL
  * @param {String} csrfToken CSRF Token
  */
-const injectJoinUs = (apps, siteUrl, csrfToken) => {
+export default (apps, siteUrl, csrfToken) => {
   // excluding `.join-us.on-nav` because it's taken care of by nav-newsletter.js
   document.querySelectorAll(`.join-us:not(.on-nav)`).forEach(element => {
     const props = element.dataset;
@@ -28,14 +28,4 @@ const injectJoinUs = (apps, siteUrl, csrfToken) => {
       })
     );
   });
-};
-
-/**
- * Inject React components
- * @param {Array} apps The existing array we are using to to track all ReactDOM.render calls
- * @param {String} siteUrl Foundation site base URL
- * @param {String} csrfToken CSRF Token
- */
-export default (apps, siteUrl, csrfToken) => {
-  injectJoinUs(apps, siteUrl, csrfToken);
 };
