+++ source/js/common/template-js-handles/index.js
@@ -1,0 +1,8 @@
+import bindFooterDonateButtonHandler from "./footer-donate-button";
+
+/**
+ * Bind event handlers
+ */
+export const bindEventHandlers = () => {
+  bindFooterDonateButtonHandler();
+};
